requirements.txt
python-telegram-bot>=20.0
qrcode[pil]>=7.4
python-barcode[images]>=0.15
Pillow>=10.0
python-dotenv>=1.0
